et ieta attempt 5

+++ Plotting/TL1et_ieta_plot.h
@@ -58,7 +58,7 @@ void TL1et_ieta_plot::InitPlots()
     
 }
 
-void TL1et_ieta_plot::Fill(const double & xVal, const double & yVal, const double & nVtxVal)
+void TL1et_ieta_plot::Fill(const double & xVal, const double & yVal, const int & nVtxVal)
 {
     //fPlot[0]->Fill(xVal,yVal);
